pyproject.toml
[project]
name = "module3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "gputil>=1.4.0",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "lightning>=2.5.1.post0",
    "matplotlib>=3.10.3",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "torch-fidelity>=0.3.0",
    "torchmetrics>=1.7.3",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
